+++ src/low.c
@@ -320,7 +320,7 @@ sock_t panet_connect(
 #ifdef COMPILE_WIN32
 DLLIMPORT
 #endif
-int panet_read(sock_t fd, void *buf, size_t len) {
+ssize_t panet_read(sock_t fd, void *buf, size_t len) {
 #ifdef COMPILE_WIN32
 	return recv(fd, (char *) buf, len, 0);
 #else
@@ -331,7 +331,7 @@ int panet_read(sock_t fd, void *buf, size_t len) {
 #ifdef COMPILE_WIN32
 DLLIMPORT
 #endif
-int panet_write(sock_t fd, const void *buf, size_t len) {
+ssize_t panet_write(sock_t fd, const void *buf, size_t len) {
 #ifdef COMPILE_WIN32
 	return send(fd, (char *) buf, len, 0);
 #else
